requirements.txt
Django==3.2.8
django-rest-framework==0.1.0
redis==3.5.3
celery==5.1.2
python-dotenv==0.19.1
django-extensions==3.1.3
PyJWT==2.3.0
random-password-generator==2.2.0